feat: booking partners stats
