+++ Examples/BookingSystem.AspNetCore.IdentityServer/Custom/Client/ClientRegistrationController.cs
@@ -50,7 +50,7 @@ namespace IdentityServer
                 registrationKey = headerValues.FirstOrDefault().Substring("Bearer ".Length);
 
             // update the booking system
-            var bookingPartner = await FakeBookingSystem.Database.GetBookingPartnerByInitialAccessToken(registrationKey);
+            var bookingPartner = await BookingPartnerTable.GetByInitialAccessToken(registrationKey);
             if (bookingPartner == null)
                 return Unauthorized("Initial Access Token is not valid, or is expired");
 
@@ -63,7 +63,7 @@ namespace IdentityServer
             bookingPartner.RedirectUris = model.RedirectUris;
             bookingPartner.Scope = model.Scope;
 
-            await FakeBookingSystem.Database.SaveBookingPartner(bookingPartner);
+            await BookingPartnerTable.Save(bookingPartner);
 
             // Read the updated client from the database and reflect back in the request
             var client = await _clients.FindClientByIdAsync(bookingPartner.ClientId);
