+++ Fakes/OpenActive.FakeDatabase.NET/Models/GrantTable.cs
@@ -1,0 +1,15 @@
+using System;
+
+namespace OpenActive.FakeDatabase.NET
+{
+    public class GrantTable
+    {
+        public string Key { get; set; }
+        public string Type { get; set; }
+        public string SubjectId { get; set; }
+        public string ClientId { get; set; }
+        public DateTime CreationTime { get; set; }
+        public DateTime? Expiration { get; set; }
+        public string Data { get; set; }
+    }
+}
